requirements.txt
alpaca-trade-api==3.1.1
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
schedule==1.2.0